+++ template/conftest.py
@@ -17,7 +17,7 @@ BACKENDS = ["speculos", "ledgercomm", "ledgerwallet"]
 
 DEVICES = ["nanos", "nanox", "nanosp", "all"]
 
-FIRMWARES = [Firmware('nanos', '2.1'), Firmware('nanox', '2.0.2'), Firmware('nanosp', '1.0.3')]
+FIRMWARES = [Firmware('nanos', '2.1'), Firmware('nanox', '2.0.2'), Firmware('nanosp', '1.0.4')]
 
 
 def pytest_addoption(parser):