+++ src/ragger/conftest/configuration.py
@@ -1,23 +1,28 @@
-from typing import Dict, Any
+from dataclasses import dataclass
 
-# The following parameters, if any, are required.
-# Please set their values in your 'conftest.py' file
-REQUIRED: Dict[str, Any] = {}
 
-# The following parameters are optional and can be used to customize your tests
-# You can set their values in your 'conftest.py' file if you want to use them
-OPTIONAL: Dict[str, Any] = {
+@dataclass
+class OptionalOptions:
+    SIDELOADED_APPS: dict
+    SIDELOADED_APPS_DIR: str
+    BACKEND_SCOPE: str
 
+
+OPTIONAL = OptionalOptions(
     # use this parameter if you want Speculos to emulate other applications being installed on the device
     # example: {"bitcoin": "Bitcoin", "ethereum": "Ethereum"}
-    "SIDELOADED_APPS": {},
+    # this would result in Speculos being launched with -lbitcoin:/path/to/bitcoin_device.elf
+    SIDELOADED_APPS=dict(),
 
     # Mandatory in case SIDELOADED_APPS is used
-    "SIDELOADED_APPS_DIR": "",
+    # Relative path to the directory that will store your libs, from the top repository of your application
+    # example: "test/lib_binaries/"
+    # place your compiled application inside with naming "name_device.elf"
+    SIDELOADED_APPS_DIR=str(),
 
     # As the backend instantiation may take some time, Ragger supports multiple backend scopes.
     # You can choose to share the backend instance between {session / module / class / function}
     # When using "session" all your tests will share a single backend instance (faster)
     # When using "function" each test will have its independent backend instance (no collusion)
-    "BACKEND_SCOPE": "class"
-}
+    BACKEND_SCOPE="class",
+)