+++ tests/conftest.py
@@ -7,7 +7,7 @@ from ragger.conftest import configuration
 # You can configure optional parameters by overriding the value of ragger.configuration.OPTIONAL_CONFIGURATION
 # Please refer to ragger/conftest/configuration.py for their descriptions and accepted values
 
-configuration.OPTIONAL["BACKEND_SCOPE"] = "function"
+configuration.OPTIONAL.BACKEND_SCOPE = "function"
 
 #########################
 ### CONFIGURATION END ###