+++ CHANGELOG.md
@@ -8,6 +8,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+
+### Changed
+
+### Fixed
+
+## [1.5.0] - 2023-02-17
+
+### Added
  - conftest: new module that provides a base_conftest for unifying conftest features
 
 ### Changed